refactor: insert Winners to Cars
- Winners 클래스를 Cars에 추가
- Winners 클래스 삭제
